+++ libraries/AD9833/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.2] - 2024-07-03
+- optimize AVR SW SPI (about factor 3.4 faster)
+- add rounding in **setPhase()** to minimize error.
+- add **float setPhaseRadians(float phase, uint8_t channel = 0)**
+- add **float getPhaseRadians(uint8_t channel = 0)**
+- add **getPowerMode()** for completeness
+- extend performance measurements (examples)
+- minor edits
+
 ## [0.4.1] - 2024-05-24
 - add support for ARDUINO_ARCH_MBED
 - add sweep example